+++ src/Sampler/RuleBased/src/ComponentProvider/SamplingRuleSpanName.php
@@ -2,39 +2,40 @@
 
 declare(strict_types=1);
 
-namespace OpenTelemetry\Contrib\Sampler\RuleBased\Configuration;
+namespace OpenTelemetry\Contrib\Sampler\RuleBased\ComponentProvider;
 
 use OpenTelemetry\Config\SDK\Configuration\ComponentProvider;
 use OpenTelemetry\Config\SDK\Configuration\ComponentProviderRegistry;
 use OpenTelemetry\Config\SDK\Configuration\Context;
+use OpenTelemetry\Config\SDK\Configuration\Validation;
 use OpenTelemetry\Contrib\Sampler\RuleBased\SamplingRule;
-use OpenTelemetry\Contrib\Sampler\RuleBased\SamplingRule\ParentRule;
+use OpenTelemetry\Contrib\Sampler\RuleBased\SamplingRule\SpanNameRule;
 use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
 
-final class SamplingRuleParent implements ComponentProvider
+final class SamplingRuleSpanName implements ComponentProvider
 {
 
     /**
      * @param array{
-     *     sampled: bool,
-     *     remote: ?bool,
+     *     pattern: string,
      * } $properties
      */
     public function createPlugin(array $properties, Context $context): SamplingRule
     {
-        return new ParentRule(
-            $properties['sampled'],
-            $properties['remote'],
+        return new SpanNameRule(
+            $properties['pattern'],
         );
     }
 
     public function getConfig(ComponentProviderRegistry $registry): ArrayNodeDefinition
     {
-        $node = new ArrayNodeDefinition('parent');
+        $node = new ArrayNodeDefinition('span_name');
         $node
             ->children()
-                ->booleanNode('sampled')->isRequired()->end()
-                ->booleanNode('remote')->defaultNull()->end()
+                ->scalarNode('pattern')
+                    ->isRequired()
+                    ->validate()->always(Validation::ensureRegexPattern())->end()
+                ->end()
             ->end()
         ;
